% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tde_baixar_dados_nome.R
\name{tde_baixar_dados_nome}
\alias{tde_baixar_dados_nome}
\title{Baixar dados de teses e dissertações com base no nome}
\usage{
tde_baixar_dados_nome(url = NULL, diretorio = ".")
}
\arguments{
\item{url}{Vetor de urls}

\item{diretorio}{Diretório onde armazenar os dados}
}
\value{
htmls
}
\description{
Baixar dados de teses e dissertações com base no nome
}
