% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tde_baixar_metadados_ano.R
\name{tde_baixar_metadados_ano}
\alias{tde_baixar_metadados_ano}
\title{Baixa tabelas com metadados das teses e dissertações}
\usage{
tde_baixar_metadados_ano(ano = NULL, diretorio = NULL)
}
\arguments{
\item{ano}{Ano no formato aaaa}

\item{diretorio}{Diretório onde armazenar os htmls}
}
\value{
html
}
\description{
Baixa tabelas com metadados das teses e dissertações
}
