% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tde_ler_dados_nomes.R
\name{tde_ler_dados_nome}
\alias{tde_ler_dados_nome}
\title{Ler teses, dissertações e livre docência}
\usage{
tde_ler_dados_nome(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Arquivos}

\item{diretorio}{Se não informar arquivos}
}
\value{
tibble
}
\description{
Ler teses, dissertações e livre docência
}
