% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tde_ler_metadados_ano.R
\name{tde_ler_metadados_ano}
\alias{tde_ler_metadados_ano}
\title{Lê tabelas de metadados das teses e dissertações}
\usage{
tde_ler_metadados_ano(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Arquivos}

\item{diretorio}{Informar se não informar arquivos}
}
\value{
tibble
}
\description{
Lê tabelas de metadados das teses e dissertações
}
